+++ src/app/helpers/tests/jest-setup.js
@@ -1,4 +1,11 @@
 import fetch from 'jest-fetch-mock'; // eslint-disable-line import/no-extraneous-dependencies
+import path from 'path';
 
 global.fetch = fetch;
-process.env.ASSETS_MANIFEST = 'mockBundleAssets.json';
+process.env.BASE_URL = 'http://localhost:7080';
+process.env.PUBLIC_DIR = 'build/public';
+process.env.PUBLIC_DIR_DEV = 'public';
+process.env.ASSETS_MANIFEST_PATH = path.resolve(
+  __dirname,
+  '../../../server/assets/fixture.json',
+);
