+++ src/server/assets/fixture.json
@@ -1,0 +1,5 @@
+{
+  "one": {
+    "js": "one.js"
+  }
+}
